requirements.txt
pywaggle[all]
requests
schedule